// ConferenciaFisica/src/app/apps/collector/estufagem-container/model/saldo-carga-marcante.model.ts
export interface SaldoCargaMarcante {
  autonumRo: number;
  autonumRcs: number;
  autonumPcs: number;
  autonumBoo: number;
  reservaCarga: string;
  numNf: string;
  autonumNf: number;
  qtde: number;
  saldo: number;
  marcante: string;
  autonumPatio: number;
  AutonumEmb: number;
  bruto: number;
  comprimento: number;
  largura: number;
  altura: number;
  autonumPro: number;
  conteiner: string;
  idMarcante: number;
}